Version 6.3.1: [ADD] Fulltext: Built-in support for Snowball, Wordnet, and Lucene stemmers! [ADD] Fulltext: Internal German stemmer added [ADD] REST: support for hierarchical collections added; details will follow [ADD] REST/GET: bind parameter added to enable external variable bindings [ADD] REST/POST: <variable/> and <output/> elements added [ADD] APIs: bind(), execute() & info() added to query iterator [ADD] APIs: add() implemented for adding documents to existing databases [MOD] REST: start was simplified; just launch basex-api/etc/basexrest(.bat) [MOD] JAX-RX: latest version was moved to our BaseX SVN repository [MOD] Warning: start class BaseXWin was renamed to BaseXGUI [MOD] Work on the upcoming Windows installer [FIX] XQuery: improved handling of extended kind tests in sequence types [FIX] XQuery 3.0 functions (e.g. format-integer()) [FIX] XQuery Update, Namespaces, collections [FIX] GUI: realtime filtering

+++ src/main/java/org/basex/data/FTMatches.java
@@ -18,13 +18,13 @@ public final class FTMatches implements Iterable<FTMatch> {
   /** Number of entries. */
   public int size;
   /** Current number of tokens. */
-  public byte sTokenNum;
+  public int sTokenNum;
 
   /**
    * Constructor.
    * @param s sets the token number
    */
-  public FTMatches(final byte s) {
+  public FTMatches(final int s) {
     reset(s);
   }
 
@@ -32,7 +32,7 @@ public final class FTMatches implements Iterable<FTMatch> {
    * Resets the match counter.
    * @param s sets the token number
    */
-  public void reset(final byte s) {
+  public void reset(final int s) {
     sTokenNum = s;
     size = 0;
   }
